+++ src/api/index.ts
@@ -2,6 +2,9 @@ const fetchQuery = async <T>(url: string): Promise<T> => {
   const response = await fetch(url);
   const data = await response.json();
 
+  if (response.status < 200 || 300 < response.status) {
+    throw new Error(data.message);
+  }
   return data;
 };
